fix(commands): forget only what the WindowServer itself forgot
Round-3 architect catch: a minimize is a windowDestroyed with
wasMinimized true whose window still exists with its memberships
— forgetting the ledger there orphans them past the quit retire.
The Events forget gates on a genuine close, and the terminated
sweep takes exactly the ids the fold removed for THAT pid (via
AppliedEffects.terminatedWindows) instead of every state-absent
id, which also matched a minimized window of another app. Two
behavior tests drive both rules through a real core with seeded
ledgers — closing the prover's stated body-gap on the terminated
sweep — in a new StickyReachCoreTests.swift (the old file hit
the ceiling); the settle slot's 800 ms carries its derivation
and single-shot statement; forward notes for #1146 recorded.
